+++ commands/serverless.js
@@ -10,6 +10,8 @@ module.exports = function (input, flags, logger, options) {
   const cwd = options.cwd
   const out = flags.out
   const env = flags.env
+  const deploymentBucket = flags.deploymentBucket
+  const executionRole = flags.executionRole
 
   if (!out) {
     return Promise.reject(new Error('"--out" is mandatory'))
@@ -19,6 +21,6 @@ module.exports = function (input, flags, logger, options) {
     .then(() => lib.applyTemplate(out)) // TODO: eventually unnecessary?
     .then(() => lib.copyWrapper(out))
     .then(() => lib.copyConfiguration(out, env))
-    .then(() => lib.registerFunctions(out, env))
+    .then(() => lib.registerFunctions(out, env, deploymentBucket, executionRole))
     .then(() => lib.registerRootProxy(out, env))
 }